Move example data to its own file
